Got TEs and rendering done.  Also finished up entities.  Last major hurdle is guns.  After that, it's just tidiying up the various methods.

+++ src/main/java/minecrafttransportsimulator/baseclasses/VehicleGroundDeviceBox.java
@@ -225,7 +225,7 @@ public class VehicleGroundDeviceBox{
 			}
 			
 			isLiquidCollidedWithGround = false;
-			for(Point3i blockPosition : liquidBox.collidingBlockPositions){
+			for(Point3d blockPosition : liquidBox.collidingBlockPositions){
 				if(!vehicle.world.isBlockLiquid(blockPosition)){
 					isLiquidCollidedWithGround = true;
 					break;
